Fixed typos, added a separate AICO trajctory example, fixed quaternion normalization, corrected node names in launch files.
