+++ exotica/src/Tools/Conversions.cpp
@@ -23,7 +23,11 @@ namespace exotica
         switch(val.rows())
         {
         case 7:
-            return KDL::Frame(KDL::Rotation::Quaternion(val(3),val(4),val(5),val(6)),KDL::Vector(val(0),val(1),val(2)));
+            {
+                double norm = val.tail(4).norm();
+                if(!(norm>0.0)) throw_pretty("Invalid quaternion!");
+                return KDL::Frame(KDL::Rotation::Quaternion(val(3)/norm, val(4)/norm, val(5)/norm, val(6)/norm), KDL::Vector(val(0),val(1),val(2)));
+            }
         case 6:
             return KDL::Frame(KDL::Rotation::RPY(val(3),val(4),val(5)),KDL::Vector(val(0),val(1),val(2)));
         case 3:
