+++ exotica_python/launch/AICOplannerTrajectory.launch
@@ -7,7 +7,7 @@
   <param name="robot_description" textfile="$(find exotica)/resources/robots/lwr_simplified.urdf" />
   <param name="robot_description_semantic" textfile="$(find exotica)/resources/robots/lwr_simplified.srdf" />
 
-  <node launch-prefix="$(arg launch_prefix)" pkg="exotica_python" type="example_distance.py" name="CollisionDistanceExampleNode" output="screen" />
+  <node launch-prefix="$(arg launch_prefix)" pkg="exotica_python" type="example_aico_trajectory.py" name="AICOPlannerPythonExampleNode" output="screen" />
 
-  <node name="rviz" pkg="rviz" type="rviz" respawn="false"	args="-d $(find exotica)/resources/rviz-distance.rviz" />
+  <node name="rviz" pkg="rviz" type="rviz" respawn="false"	args="-d $(find exotica)/resources/rviz.rviz" />
 </launch>
