+++ exotica_python/scripts/example_aico_trajectory.py
@@ -2,22 +2,16 @@
 
 import pyexotica as exo
 from pyexotica.publish_trajectory import *
-from numpy import array
-import math
-
-def figureEight(t):
-    return array([0.0, math.sin(t * 2.0 * math.pi * 0.5) * 0.1, math.sin(t * math.pi * 0.5) * 0.2, 0.0, 0.0, 0.0])
 
 exo.Setup.initRos()
-solver=exo.Setup.loadSolver('{exotica}/resources/configs/aico_solver_demo_eight.xml')
+solver=exo.Setup.loadSolver('{exotica}/resources/configs/aico_trajectory.xml')
 problem = solver.getProblem()
 
 for t in range(0,problem.T):
-    if t<problem.T/5:
+    if float(t)*problem.tau<0.8:
         problem.setRho('Frame',0.0,t)
     else:
         problem.setRho('Frame',1e5,t)
-        problem.setGoal('Frame',figureEight(t*problem.tau),t)
 
 solution = solver.solve()
